+++ esp_cam_sensor/sensors/sc132gs/private_include/sc132gs_settings.h
@@ -1,0 +1,24 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#pragma once
+
+#include <stdio.h>
+#include <stdint.h>
+#include <sdkconfig.h>
+#include "sc132gs_regs.h"
+#include "sc132gs_types.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+#include "sc132gs_mipi_2lane_24Minput_544x640_raw10_60fps.h"
+
+#include "sc132gs_mipi_2lane_24Minput_1080x1280_raw10_60fps.h"
+
+#ifdef __cplusplus
+}
+#endif