+++ esp_cam_sensor/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added SC2337P MIPI RAW sensor driver.
 - Added BF3031 DVP Global color sensor.
+- Added SC132GS MIPI driver.
 
 ## 2.4.0
 - Added SC2331 MIPI RAW sensor driver.